services/kafka-to-questdb/requirements.txt
kafka-python
pandas